+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package rviz_satellite
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Forthcoming
+-----------
+* Rewrite of the tile loader
+* Display an error if the tile server is invalid
+* Move map towards the robot's z-position
+* Use XDG cache path for caching tiles
+* Incorporate orientation of tile frame when rendering map
+
 1.1.0 (2018-12-04)
 ------------------
 * Integrate package version via CMake